requirements.txt
python-etcd3>=0.13.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
requests>=2.30.0
psycopg2-binary>=2.9.5
sqlalchemy>=2.0.0
pyyaml>=6.0
Jinja2
python-multipart
uuid
pydantic